requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-decouple==2.1
djangorestframework==3.11.1
idna==2.10
install==1.3.3
numpy==1.19.1
pandas==1.1.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
